Minor file fixes
